Move pickCertificationChallenges to certificationChallengesService
