+++ api/lib/domain/services/placement-profile-service.js
@@ -1,14 +1,10 @@
 const _ = require('lodash');
 const bluebird = require('bluebird');
 
-const KnowledgeElement = require('../models/KnowledgeElement');
 const UserCompetence = require('../models/UserCompetence');
-const Challenge = require('../models/Challenge');
 const PlacementProfile = require('../models/PlacementProfile');
 const assessmentRepository = require('../../infrastructure/repositories/assessment-repository');
 const assessmentResultRepository = require('../../infrastructure/repositories/assessment-result-repository');
-const challengeRepository = require('../../infrastructure/repositories/challenge-repository');
-const answerRepository = require('../../infrastructure/repositories/answer-repository');
 const knowledgeElementRepository = require('../../infrastructure/repositories/knowledge-element-repository');
 const competenceRepository = require('../../infrastructure/repositories/competence-repository');
 const scoringService = require('./scoring/scoring-service');
@@ -19,68 +15,6 @@ async function getPlacementProfile({ userId, limitDate, isV2Certification = true
     return _generatePlacementProfileV2({ userId, profileDate: limitDate, competences, allowExcessPixAndLevels });
   }
   return _generatePlacementProfileV1({ userId, profileDate: limitDate, competences });
-}
-
-//TODO: Est-ce une méthode du placement profile ? (ou de CertificationChallengeService ?)
-async function pickCertificationChallenges(placementProfile) {
-  const knowledgeElementsByCompetence = await knowledgeElementRepository
-    .findUniqByUserIdGroupedByCompetenceId({ userId: placementProfile.userId, limitDate: placementProfile.profileDate });
-
-  const knowledgeElements = KnowledgeElement.findDirectlyValidatedFromGroups(knowledgeElementsByCompetence);
-  const answerIds = _.map(knowledgeElements, 'answerId');
-
-  const challengeIdsCorrectlyAnswered = await answerRepository.findChallengeIdsFromAnswerIds(answerIds);
-
-  const allChallenges = await challengeRepository.findOperative();
-  const challengesAlreadyAnswered = challengeIdsCorrectlyAnswered.map((challengeId) => Challenge.findById(allChallenges, challengeId));
-
-  challengesAlreadyAnswered.forEach((challenge) => {
-    if (!challenge) {
-      return;
-    }
-
-    const userCompetence = _getUserCompetenceByChallengeCompetenceId(placementProfile.userCompetences, challenge);
-
-    if (!userCompetence || !userCompetence.isCertifiable()) {
-      return;
-    }
-
-    challenge.skills
-      .filter((skill) => _skillHasAtLeastOneChallenge(skill, allChallenges))
-      .forEach((publishedSkill) => userCompetence.addSkill(publishedSkill));
-  });
-
-  const userCompetences = UserCompetence.orderSkillsOfCompetenceByDifficulty(placementProfile.userCompetences);
-
-  userCompetences.forEach((userCompetence) => {
-    const testedSkills = [];
-    userCompetence.skills.forEach((skill) => {
-      if (!userCompetence.hasEnoughChallenges()) {
-        const challengesToValidateCurrentSkill = Challenge.findBySkill({ challenges: allChallenges, skill });
-        const challengesLeftToAnswer = _.difference(challengesToValidateCurrentSkill, challengesAlreadyAnswered);
-
-        const challengesPoolToPickChallengeFrom = (_.isEmpty(challengesLeftToAnswer)) ? challengesToValidateCurrentSkill : challengesLeftToAnswer;
-        const challenge = _.sample(challengesPoolToPickChallengeFrom);
-
-        challenge.testedSkill = skill;
-        testedSkills.push(skill);
-
-        userCompetence.addChallenge(challenge);
-      }
-    });
-    userCompetence.skills = testedSkills;
-  });
-
-  return _.flatMap(userCompetences, (userCompetence) => userCompetence.challenges);
-}
-
-function _getUserCompetenceByChallengeCompetenceId(userCompetences, challenge) {
-  return challenge ? userCompetences.find((userCompetence) => userCompetence.id === challenge.competenceId) : null;
-}
-
-function _skillHasAtLeastOneChallenge(skill, challenges) {
-  const challengesBySkill = Challenge.findBySkill({ challenges, skill });
-  return challengesBySkill.length > 0;
 }
 
 async function _createUserCompetencesV1({ allCompetences, userLastAssessments, limitDate }) {
@@ -183,5 +117,4 @@ async function getPlacementProfilesWithSnapshotting({ userIdsAndDates, competenc
 module.exports = {
   getPlacementProfile,
   getPlacementProfilesWithSnapshotting,
-  pickCertificationChallenges,
 };
